+++ lib/packages/core/lib/src/views/media/media_preview.dart
@@ -73,7 +73,7 @@ class _LMChatMediaPreviewScreenState extends State<LMChatMediaPreviewScreen> {
   @override
   Widget build(BuildContext context) {
     return ValueListenableBuilder(
-        valueListenable: LMChatTheme.themeNotifierBloc,
+        valueListenable: LMChatTheme.themeNotifier,
         builder: (context, _, child) {
           return _screenBuilder.scaffold(
             onPopInvoked: (p0) {